Move struct Token into a separate file
