+++ src/artm/core/check_messages.h
@@ -14,6 +14,7 @@
 #include "artm/core/common.h"
 #include "artm/core/exceptions.h"
 #include "artm/core/internals.pb.h"
+#include "artm/core/token.h"
 
 namespace artm {
 namespace core {
