+++ src/artm/core/dictionary.h
@@ -11,6 +11,7 @@
 #include "artm/messages.pb.h"
 #include "artm/core/common.h"
 #include "artm/core/thread_safe_holder.h"
+#include "artm/core/token.h"
 
 namespace artm {
 namespace core {
