+++ src/artm/core/phi_matrix.h
@@ -11,6 +11,7 @@
 
 #include "artm/core/common.h"
 #include "artm/core/internals.pb.h"
+#include "artm/core/token.h"
 
 namespace artm {
 namespace core {
