+++ src/artm/core/token.cc
@@ -1,0 +1,9 @@
+// Copyright 2016, Additive Regularization of Topic Models.
+
+#include "artm/core/token.h"
+
+namespace artm {
+namespace core {
+
+}  // namespace core
+}  // namespace artm
