+++ src/artm/score_calculator_interface.h
@@ -12,6 +12,7 @@
 #include "artm/core/common.h"
 #include "artm/core/phi_matrix.h"
 #include "artm/core/exceptions.h"
+#include "artm/core/token.h"
 #include "artm/messages.pb.h"
 
 namespace artm {
