+++ src/artm_tests/repeatable_result_test.cc
@@ -9,6 +9,7 @@
 #include "artm/core/common.h"
 #include "artm/core/exceptions.h"
 #include "artm/core/helpers.h"
+#include "artm/core/token.h"
 #include "artm/messages.pb.h"
 
 #include "artm/core/internals.pb.h"
